+++ python/ray/tune/trainable.py
@@ -124,8 +124,13 @@ class Trainable(object):
 
     @classmethod
     def resource_help(cls, config):
-        """Returns a help string for configuring this trainable's resources."""
+        """
+        Args:
+            config (dict): The Trainer's config dict.
 
+        Returns:
+             str: A help string for configuring this trainable's resources.
+        """
         return ""
 
     def current_ip(self):
