+++ rllib/agents/pg/pg_tf_policy.py
@@ -12,7 +12,7 @@ from ray.rllib.utils import try_import_tf
 tf = try_import_tf()
 
 
-def tf_pg_loss(policy, model, dist_class, train_batch):
+def pg_tf_loss(policy, model, dist_class, train_batch):
     """The basic policy gradients loss."""
     logits, _ = model.from_batch(train_batch)
     action_dist = dist_class(logits, model)
@@ -31,4 +31,4 @@ PGTFPolicy = build_tf_policy(
     name="PGTFPolicy",
     get_default_config=lambda: ray.rllib.agents.pg.pg.DEFAULT_CONFIG,
     postprocess_fn=post_process_advantages,
-    loss_fn=tf_pg_loss)
+    loss_fn=pg_tf_loss)
