+++ rllib/agents/pg/pg_torch_policy.py
@@ -3,7 +3,7 @@ from __future__ import division
 from __future__ import print_function
 
 import ray
-from ray.rllib.agents.pg.tf_pg_policy import post_process_advantages
+from ray.rllib.agents.pg.pg_tf_policy import post_process_advantages
 from ray.rllib.evaluation.postprocessing import Postprocessing
 from ray.rllib.policy.sample_batch import SampleBatch
 from ray.rllib.policy.torch_policy_template import build_torch_policy
@@ -12,7 +12,7 @@ from ray.rllib.utils.framework import try_import_torch
 torch, _ = try_import_torch()
 
 
-def torch_pg_loss(policy, model, dist_class, train_batch):
+def pg_torch_loss(policy, model, dist_class, train_batch):
     """The basic policy gradients loss."""
     logits, _ = model.from_batch(train_batch)
     action_dist = dist_class(logits, model)
@@ -34,6 +34,6 @@ def pg_loss_stats(policy, train_batch):
 PGTorchPolicy = build_torch_policy(
     name="PGTorchPolicy",
     get_default_config=lambda: ray.rllib.agents.pg.pg.DEFAULT_CONFIG,
-    loss_fn=torch_pg_loss,
+    loss_fn=pg_torch_loss,
     stats_fn=pg_loss_stats,
     postprocess_fn=post_process_advantages)
