+++ rllib/tests/test_external_multi_agent_env.py
@@ -8,7 +8,7 @@ import random
 import unittest
 
 import ray
-from ray.rllib.agents.pg.tf_pg_policy import PGTFPolicy
+from ray.rllib.agents.pg.pg_tf_policy import PGTFPolicy
 from ray.rllib.optimizers import SyncSamplesOptimizer
 from ray.rllib.evaluation.rollout_worker import RolloutWorker
 from ray.rllib.evaluation.worker_set import WorkerSet
